% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/JPR_multiple.R
\docType{package}
\name{JPR_multiple}
\alias{JPR_multiple}
\alias{JPR_multiple-package}
\title{Measuring Japanese readability (Multiple ver.)}
\usage{
JPR_multiple(file.df)
}
\arguments{
\item{file.df}{information of text file. The data frame must have three columns. The First column is id, the second is description, the third is filename(include path)}

\item{result.only}{Show only results without likihood(default is FALSE)}
}
\value{
data.frame
}
\description{
Measuring Japanese readability (Multiple ver.)
}
\examples{
setwd("/User/Username/Documents/")
sample.df <- data.frame(id = c(1:3), text  = c("Abe Shinzo", "Kan Naoto", "Ozawa Ichiro"), file_path = c("AS.txt", "KN.txt", "OI.txt"))
JPR_multiple(sample.df)
}
\author{
Jaehyun SONG
}

