% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/JPreadable.R
\docType{package}
\name{JPreadable}
\alias{JPreadable}
\alias{JPreadable-package}
\title{Measuring Japanese readability}
\usage{
JPreadable(file)
}
\arguments{
\item{file}{a name of text file(include its path)}
}
\value{
List type(Raw text, Liklihood table, Grades with maximum liklihood, Median)
}
\description{
Measuring Japanese readability
}
\examples{
setwd("/User/Username/Documents/")
JPreadble("sample_text.txt")
}
\author{
Jaehyun SONG
}
\references{
Satoshi Sato, Suguru Matsuyoshi and Yohsuke Kondoh. 2008. Automatic Assessment of Japanese Text Readability Based on a Textbook Corpus. Proceedings of the Sixth International Language Resources and Evaluation (LREC’08). European Language Resources Association (ELRA). Marrakech, Morocco
}

