% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/JPreadable.R
\docType{package}
\name{JPreadable}
\alias{JPreadable}
\alias{JPreadable-package}
\title{Measuring Japanese readability}
\usage{
JPreadable(file)
}
\arguments{
\item{file}{a name of text file(include its path)}
}
\value{
List type
}
\description{
Measuring Japanese readability
}
\examples{
setwd("/User/Username/Documents/")
JPreadble("sample_text.txt")
}
\author{
Jaehyun SONG
}

